Refactor, including file renaming and more functions added to the functions file
